requirements.txt
fastapi
uvicorn
pyserial
motor
pymongo
firebase-admin
python-dotenv
websockets
pydantic
requests